requirements.txt
langchain
langchain-community
langchain-core
faiss-cpu==1.8.0
PyPDF2==3.0.1
python-dotenv==1.0.1
langchain-google-genai
sentence-transformers
streamlit
flask
streamlit-chat